requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.0  # LoRA/QLoRA
bitsandbytes>=0.41.0  # Quantization
accelerate>=0.25.0
trl>=0.7.0  # Training utilities

# Experiment Tracking
mlflow>=2.9.0
wandb>=0.16.0

# Serving & Inference
vllm>=0.2.0  # Fast inference
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlparse>=0.4.4  # For SQL parsing/formatting

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Database & Storage
sqlalchemy>=2.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Cloud & Deployment (optional)
# boto3>=1.29.0  # AWS
# google-cloud-storage>=2.10.0  # GCP
